+++ OASIS Omniverse/STARAPIClient/Jobs/PumpActionKind.cs
@@ -1,3 +1,0 @@
-﻿namespace NextGenSoftware.OASIS.STARAPI.Client;
-
-internal enum PumpActionKind { Auth, Inventory, SendItem, UseItem, AddItemLog, OperationCallback }